Reference

Index

TidierPlots.aes

TidierPlots.alpha_scale_to_ggoptions

TidierPlots.facet_grid

TidierPlots.geom_bar

TidierPlots.geom_boxplot

TidierPlots.geom_col

TidierPlots.geom_contour

TidierPlots.geom_contour_filled

TidierPlots.geom_density

TidierPlots.geom_density2d

TidierPlots.geom_density2d_filled

TidierPlots.geom_density_2d

TidierPlots.geom_density_2d_filled

TidierPlots.geom_errorbar

TidierPlots.geom_errorbarh

TidierPlots.geom_histogram

TidierPlots.geom_hline

TidierPlots.geom_jitter

TidierPlots.geom_label

TidierPlots.geom_line

TidierPlots.geom_linerange

TidierPlots.geom_path

TidierPlots.geom_point

TidierPlots.geom_pointrange

TidierPlots.geom_rainclouds

TidierPlots.geom_segment

TidierPlots.geom_smooth

TidierPlots.geom_step

TidierPlots.geom_text

TidierPlots.geom_tile

TidierPlots.geom_violin

TidierPlots.geom_vline

TidierPlots.guides

TidierPlots.handle_area

TidierPlots.handle_point_color_and_fill

TidierPlots.handle_polygon

TidierPlots.handle_rect

TidierPlots.handle_ribbon

TidierPlots.label_bytes

TidierPlots.label_currency

TidierPlots.label_date

TidierPlots.label_log

TidierPlots.label_number

TidierPlots.label_ordinal

TidierPlots.label_percent

TidierPlots.label_pvalue

TidierPlots.label_scientific

TidierPlots.label_wrap

TidierPlots.linewidth_scale_to_ggoptions

TidierPlots.make_alpha_lookup_continuous

TidierPlots.make_linewidth_lookup_continuous

TidierPlots.position_facets

TidierPlots.scale_custom

TidierPlots.shape_scale_to_ggoptions

TidierPlots.size_scale_to_ggoptions

TidierPlots.update_legend

Reference - Exported functions

TidierPlots.aes — Method

aes(args...; kwargs...)

Details

TBD

TidierPlots.facet_grid — Method

facet_grid(;rows, cols, scales = "fixed", switch = "none")

facetgrid() forms a matrix of panels defined by row and column faceting variables. It is most useful when you have two discrete variables, and all combinations of the variables exist in the data. If you have only one variable with many levels, try facetwrap().

Arguments

rows (required): Variable to use for the rows of the matrix

cols (required): Variable to use for the columns of the matrix

scales (optional): Should the scales be fixed or free? Options: "free", "freex", "freey"

switch (optional): Flip the labels from their default "top and right". Options: "x", "y", "both".

TidierPlots.geom_bar — Function

geom_bar(aes(...), ...)
geom_bar(plot::GGPlot, aes(...), ...)

Represent the counts of a grouping variable as columns.

Details

The columns are stacked by default, and the behavior can be changed with the "position" argument. The position can either be "stack" or "dodge". If the argument is "dodge", then a a grouping variable will also need to be supplied to aes. Alternatively you can supply the grouping variable to dodge within the aesthetic.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.BarPlot)

Required Aesthetics

x OR y (not both)

Optional Aesthetics (see aes)

color / colour

strokecolor / strokecolour

dodge

group

Optional Arguments

position: "stack" (default) or "dodge"

stroke / strokewidth

strokecolor / strokecolour

direction: :y (default) or :x

dodge_gap

gap

Examples

penguins = penguins_data();
# vertical bar plot
ggplot(penguins) + geom_bar(@aes(x = species))

# horizontal bar plot
ggplot(penguins) + geom_bar(@aes(y = species))

# stacked
ggplot(penguins, @aes(x = species, fill=sex)) + geom_bar()

# dodged
ggplot(penguins, @aes(x = species, dodge = sex)) + geom_bar()

TidierPlots.geom_boxplot — Function

geom_boxplot(aes(...), ...)
geom_boxplot(plot::GGPlot, aes(...), ...)

Compactly displays the distribution of continuous data.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.BoxPlot)

Required Aesthetics

x (integer or categorical)

y (numeric)

Optional Aesthetics (see aes)

color / colour (used in conjunction with dodge)

dodge

Optional Arguments

orientation=:vertical: orientation of box (:vertical or :horizontal)

width=1

gap=0.2

show_notch=false

nothchwidth=0.5

show_median=true

dodge_gap=0.03

Examples

penguins = penguins_data();
ggplot(penguins, @aes(x=species, y=bill_length_mm)) +
    geom_boxplot()

ggplot(penguins, @aes(y=species, x=bill_length_mm)) +
    geom_boxplot()

ggplot(penguins, @aes(x=species, y=bill_length_mm, fill=sex)) +
    geom_boxplot()

TidierPlots.geom_col — Function

geom_col(aes(...), ...)
geom_col(plot::GGPlot, aes(...), ...)

Represent data as columns.

Details

The columns are stacked by default, and the behavior can be changed with the "position" argument. The position can either be "stack" or "dodge". If the argument is "dodge", then a a grouping variable will also need to be supplied to aes. Alternatively you can supply the grouping variable to dodge within the aesthetic.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.BarPlot)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

strokecolor / strokecolour

dodge

group

Optional Arguments

position: "stack" (default) or "dodge"

stroke / strokewidth

strokecolor / strokecolour

direction: :y (default) or :x

dodge_gap

gap

Examples

df = @chain penguins begin
    @group_by(species, sex)
    @summarize(mean_bill_length_mm = mean(bill_length_mm))
    @ungroup()
end

ggplot(df) +
    geom_col(@aes(x = species, y = mean_bill_length_mm))

# dodge using the group and position arguments
ggplot(df) +
    geom_col(@aes(x = species, y = mean_bill_length_mm, group = sex),
             position="dodge")

# dodge using the dodge aesthetic
ggplot(df) +
    geom_col(@aes(x = species, y = mean_bill_length_mm, dodge = sex))

# color based on grouping variable
ggplot(df) +
    geom_col(@aes(x = species, y = mean_bill_length_mm, color = sex))

TidierPlots.geom_contour — Function

geom_density_2d(aes(...), ...)
geom_density_2d(plot::GGPlot, aes(...), ...)

geom_density2d(aes(...), ...)
geom_density2d(plot::GGPlot, aes(...), ...)

geom_contour(aes(...), ...)
geom_contour(plot::GGPlot, aes(...), ...)

Represent data as a 2D density contour using 2D kernel density estimation.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

strokewidth / stroke

linestyle / linetype

alpha

group

Optional Arguments

breaks / levels

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d()

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d(color = :black, stroke = 2)

See also geom_contour_filled and geom_density_2d_filled for similar geoms.

TidierPlots.geom_contour_filled — Function

geom_density_2d_filled(aes(...), ...)
geom_density_2d_filled(plot::GGPlot, aes(...), ...)

geom_density2d_filled(aes(...), ...)
geom_density2d_filled(plot::GGPlot, aes(...), ...)

geom_contour_filled(aes(...), ...)
geom_contour_filled(plot::GGPlot, aes(...), ...)

Represent data as a filled 2D density contour using 2D kernel density estimation.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

alpha

group

Optional Arguments

breaks / levels

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d_filled()

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d_filled(color = :black, stroke = 2)

See also geom_contour and geom_density_2d for similar geoms.

TidierPlots.geom_density — Function

geom_density(aes(...), ...)
geom_density(plot::GGPlot, aes(...), ...)

Represent data as a smooth density curve.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

Optional Aesthetics (see aes)

NA

Optional Arguments

color / colour

colormap / palette

strokecolor / strokecolour

strokewidth / stroke

linestyle / linetype

direction=:x

npoints=200

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm)) + geom_density()

ggplot(penguins, @aes(x=bill_length_mm)) +
    geom_density(color = :black, stroke = 2)

TidierPlots.geom_density2d — Function

geom_density_2d(aes(...), ...)
geom_density_2d(plot::GGPlot, aes(...), ...)

geom_density2d(aes(...), ...)
geom_density2d(plot::GGPlot, aes(...), ...)

geom_contour(aes(...), ...)
geom_contour(plot::GGPlot, aes(...), ...)

Represent data as a 2D density contour using 2D kernel density estimation.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

strokewidth / stroke

linestyle / linetype

alpha

group

Optional Arguments

breaks / levels

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d()

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d(color = :black, stroke = 2)

See also geom_contour_filled and geom_density_2d_filled for similar geoms.

TidierPlots.geom_density2d_filled — Function

geom_density_2d_filled(aes(...), ...)
geom_density_2d_filled(plot::GGPlot, aes(...), ...)

geom_density2d_filled(aes(...), ...)
geom_density2d_filled(plot::GGPlot, aes(...), ...)

geom_contour_filled(aes(...), ...)
geom_contour_filled(plot::GGPlot, aes(...), ...)

Represent data as a filled 2D density contour using 2D kernel density estimation.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

alpha

group

Optional Arguments

breaks / levels

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d_filled()

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d_filled(color = :black, stroke = 2)

See also geom_contour and geom_density_2d for similar geoms.

TidierPlots.geom_density_2d — Function

geom_density_2d(aes(...), ...)
geom_density_2d(plot::GGPlot, aes(...), ...)

geom_density2d(aes(...), ...)
geom_density2d(plot::GGPlot, aes(...), ...)

geom_contour(aes(...), ...)
geom_contour(plot::GGPlot, aes(...), ...)

Represent data as a 2D density contour using 2D kernel density estimation.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

strokewidth / stroke

linestyle / linetype

alpha

group

Optional Arguments

breaks / levels

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d()

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d(color = :black, stroke = 2)

See also geom_contour_filled and geom_density_2d_filled for similar geoms.

TidierPlots.geom_density_2d_filled — Function

geom_density_2d_filled(aes(...), ...)
geom_density_2d_filled(plot::GGPlot, aes(...), ...)

geom_density2d_filled(aes(...), ...)
geom_density2d_filled(plot::GGPlot, aes(...), ...)

geom_contour_filled(aes(...), ...)
geom_contour_filled(plot::GGPlot, aes(...), ...)

Represent data as a filled 2D density contour using 2D kernel density estimation.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Density)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

alpha

group

Optional Arguments

breaks / levels

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d_filled()

ggplot(penguins, @aes(x=bill_length_mm, y=bill_depth_mm)) +
    geom_density_2d_filled(color = :black, stroke = 2)

See also geom_contour and geom_density_2d for similar geoms.

TidierPlots.geom_errorbar — Method

geom_errorbar(aes(...), ...)
geom_errorbar(plot::GGPlot, aes(...), ...)

Represents data as a vertical interval.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Rangebars)

Required Aesthetics

x

ymin

ymax

Optional Aesthetics (see aes)

NA

Optional Arguments

color / colour

direction=:y

linewidth

whiskerwidth / width

Examples

df = DataFrame(
    trt   = [1, 1, 2, 2],
    resp  = [1, 5, 3, 4],
    group = [1, 2, 1, 2],
    lower = [0.8, 4.6, 2.4, 3.6],
    upper = [1.1, 5.3, 3.3, 4.2],
)

ggplot(df, @aes(x = trt, ymin = lower, ymax = upper)) +
    geom_errorbar(width=20, linewidth=2)

TidierPlots.geom_errorbarh — Method

geom_errorbarh(aes(...), ...)
geom_errorbarh(plot::GGPlot, aes(...), ...)

Represents data as a horizontal interval.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Rangebars)

Required Aesthetics

y

xmin

xmax

Optional Aesthetics (see aes)

NA

Optional Arguments

color / colour

direction=:x

linewidth

whiskerwidth / width

Examples

df = DataFrame(
    trt   = [1, 1, 2, 2],
    resp  = [1, 5, 3, 4],
    group = [1, 2, 1, 2],
    lower = [0.8, 4.6, 2.4, 3.6],
    upper = [1.1, 5.3, 3.3, 4.2],
)

ggplot(df, @aes(y = trt, xmin = lower, xmax = upper)) +
    geom_errorbarh(linewidth=2)

TidierPlots.geom_histogram — Function

geom_histogram(aes(...), ...)
geom_histogram(plot::GGPlot, aes(...), ...)

Represents data as a histogram.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Hist)

Required Aesthetics

x

Optional Aesthetics (see aes)

NA

Optional Arguments

bins: number of bins or a vector of bin edges

binwidth: width of each bin (alternative to bins)

center: value that should be at the center of a bin (useful for discrete data)

boundary: value that should be at the edge of a bin

normalization

color / colour

stroke / strokewidth

strokecolor / strokecolour

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x = bill_length_mm)) +
    geom_histogram()

ggplot(penguins, @aes(x = bill_length_mm)) +
    geom_histogram(normalization=:probability, bins=20)

# For discrete/integer data, use center to align bars with tick marks
# This centers a bin on 0, so tick marks appear at bar centers
ggplot(df, @aes(x = counts)) +
    geom_histogram(binwidth=1, center=0)

TidierPlots.geom_hline — Method

geom_hline(aes(...), ...)
geom_hline(plot::GGPlot, aes(...), ...)

Plot a horizontal line at the given y-intercept(s).

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.HLines)

Required Aesthetics

NA

Optional Aesthetics (see aes)

yintercept(s)

color / colour

Optional Arguments

yintercept(s)

color / colour

linewidth

linestyle / linetype

colormap / palette

alpha

Examples

# Plot only a single y-intercept
ggplot() + geom_hline(yintercept = 3)

# Plot multiple y-intercepts
ggplot() + geom_hline(yintercept = [-1, 4])

# Plot multiple y-intercepts mapped to a column
df = DataFrame(y = rand(4))
ggplot(df, @aes(yintercept = y)) + geom_hline()

TidierPlots.geom_jitter — Function

geom_jitter(aes(...); width = 0.4, height = 0.4, …)

Scatter plot with random position noise to mitigate over-plotting.

Required aesthetics

x

y

TidierPlots.geom_label — Function

geom_label(aes(...), ...)
geom_label(plot::GGPlot, aes(...), ...)

Plot text on a graph.

ggplot2 deviation

Currently this method is the same as geom_text, and does not put the text in a rectangle like in ggplot2.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Text)

Required Aesthetics

x

y

text

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

colormap / palette

align: tuple of positions (e.g. (:left, :bottom))

font

justification

rotation

fontsize

strokewidth

strokecolor

glowwidth / glow

glowcolor / glowcolour

word_wrap_width

alpha

Examples

df = DataFrame(
    x = [1,1,2,2],
    y = [1,2,1,2],
    t = ["A", "B", "C", "D"]
)

ggplot(df, @aes(x=x, y=y, text=t, color=t)) + geom_label()

ggplot(df, @aes(x=x, y=y, color=t)) +
    geom_label(@aes(text=t), fontsize=24, align=(:left, :bottom), font=:bold) +
    geom_point() +
    lims(x = (0, 3), y = (0, 3))

TidierPlots.geom_line — Function

geom_line(aes(...), ...)
geom_line(plot::GGPlot, aes(...), ...)

Represents data as connected points in the order of the variable on the x-axis.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Lines)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

colormap / palette

linestyle / linetype

linewidth

alpha

Examples

xs = range(0, 2pi, length=30)
df = DataFrame(x = xs, y = sin.(xs))

ggplot(df, @aes(x = x, y = y)) + geom_line()

TidierPlots.geom_linerange — Function

geom_linerange(aes(...), ...)
geom_linerange(plot::GGPlot, aes(...), ...)

Displays a vertical line from ymin to ymax (without whiskers).

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Rangebars)

Required Aesthetics

x

ymin

ymax

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

linewidth

alpha

Examples

df = DataFrame(
    x = [1, 2, 3, 4],
    ymin = [1, 2, 1.5, 2.5],
    ymax = [3, 4, 3.5, 4.5]
)

ggplot(df, @aes(x = x, ymin = ymin, ymax = ymax)) + geom_linerange()

TidierPlots.geom_path — Function

geom_path(aes(...), ...)
geom_path(plot::GGPlot, aes(...), ...)

Represents data as connected points in the order in which they appear in the data.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Lines)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

colormap / palette

linestyle / linetype

linewidth

alpha

Examples

ggplot(penguins, @aes(x = bill_length_mm, y = bill_depth_mm)) +
    geom_path()

TidierPlots.geom_point — Function

geom_point(aes(...), ...)
geom_point(plot::GGPlot, aes(...), ...)

The point geom is used to create scatterplots. The scatterplot is most useful for displaying the relationship between two continuous variables. It can be used to compare one continuous and one categorical variable, or two categorical variables, but other charts are usually more appropriate. A bubblechart is a scatterplot with a third variable mapped to the size of points.

Arguments

plot::GGPlot (optional): a plot object to add this geom to. This is typically used to facilitate creating your ggplot as part of a @chain.

data (DataFrame): Data to use for this geom. If not provided, the geom will inherit the data from ggplot.

aes(...): the names of the columns in the DataFrame that will be used in the mapping

inherit_aes: should the geom inherit aes from the ggplot?

...: options that are not mapped to a column (passed to Makie.Scatter)

Overplotting

The biggest potential problem with a scatterplot is overplotting: whenever you have more than a few points, points may be plotted on top of one another. This can severely distort the visual appearance of the plot. There is no one solution to this problem, but there are some techniques that can help. You can add additional information with geomsmooth(), or if you have few unique x values, geomboxplot() may also be useful. Another technique is to make the points transparent (e.g. geompoint(alpha = 0.05)) or very small (e.g. geompoint(shape = '.')).

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

fill

shape

size

stroke

Optional Arguments

color / colour

colormap / palette

marker / shape

markersize / size

strokewidth / stroke

strokecolor / strokecolour

glowwidth / glow

glowcolor / glowcolour

alpha

Examples

penguins = penguins_data();

p = ggplot(penguins, @aes(bill_length_mm, bill_depth_mm))
p + geom_point()

# add aesthetic mappings
p + geom_point(aes(colour = :sex))

TidierPlots.geom_pointrange — Method

geom_pointrange(aes(...), ...)
geom_pointrange(plot::GGPlot, aes(...), ...)

Displays a point with a vertical line from ymin to ymax.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column

Required Aesthetics

x

y

ymin

ymax

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

linewidth

size / markersize

alpha

Examples

df = DataFrame(
    x = [1, 2, 3, 4],
    y = [2, 3, 2.5, 3.5],
    ymin = [1, 2, 1.5, 2.5],
    ymax = [3, 4, 3.5, 4.5]
)

ggplot(df, @aes(x = x, y = y, ymin = ymin, ymax = ymax)) + geom_pointrange()

TidierPlots.geom_rainclouds — Function

geom_rainclouds(aes(...), ...)
geom_rainclouds(plot::GGPlot, aes(...), ...)

"Raincloud" plots are a combination of a (half) violin plot, box plot and scatter plots. The three together can make an appealing and informative visual, particularly for large N datasets.

Arguments

plot::GGPlot (optional): a plot object to add this geom to. This is typically used to facilitate creating your ggplot as part of a @chain.

data (DataFrame): Data to use for this geom. If not provided, the geom will inherit the data from ggplot.

aes(...): the names of the columns in the DataFrame that will be used in the mapping

inherit_aes: should the geom inherit aes from the ggplot?

...: options that are not mapped to a column (passed to Makie.RainClouds)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

size

stroke

dodge

Optional Arguments

boxplot_nudge: Default 0.075. Determines the distance away the boxplot should be placed from the center line when center_boxplot is false. This is the value used to recentering the boxplot.

boxplot_width: Default 0.1. Determines the width of the boxplot.

center_boxplot: Default true. Determines whether the boxplot should be centered on the category.

cloud_width: Default 0.75. Determines the width of the violin.

clouds: Default violin. Valid values: [violin, hist, nothing]clouds

color / colour: Default will be blue unless changed by palette or aes.

dodge_gap: Default 0.01. Determines the gap between the dodged boxes.

gap: Default 0.2. Distance between the main elements of the plot.

hist_bins: Default 30. If clouds == hist, determines the number of bins for the histogram.

markersize / size: Default 2. Size of markers in the scatterplot.

plot_boxplots: Default true. Determines whether the boxplot should be plotted.

show_boxplot_outliers: Default false. Determines whether the outliers should be shown in the boxplot.

show_median: Default true. Determines whether the median should be shown in the boxplot.

side: Default :left. Determines the side of the violin relative to the points.

strokewidth / stroke: Default 1. Width of the stroke around the boxplot.

whisker_width: Default 0.5. Width of the whiskers in the boxplot.

Examples

penguins = penguins_data();

ggplot(penguins) +
    geom_rainclouds(@aes(x = species, y = bill_depth_mm/10, color = species), size = 5) +
    scale_y_continuous(labels = "{:.1f} cm") +
    labs(title = "Bill Depth by Species", x = "Species", y = "Bill Depth") +
    theme_minimal()

TidierPlots.geom_segment — Function

geom_segment(aes(...), ...)
geom_segment(plot::GGPlot, aes(...), ...)

Draws line segments between pairs of points.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.LineSegments)

Required Aesthetics

x: x-coordinate of segment start

y: y-coordinate of segment start

xend: x-coordinate of segment end

yend: y-coordinate of segment end

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

colormap / palette

linestyle / linetype

linewidth

alpha

Examples

df = DataFrame(
    x = [1, 2, 3],
    y = [1, 2, 1],
    xend = [2, 3, 4],
    yend = [2, 1, 2]
)

ggplot(df, @aes(x = x, y = y, xend = xend, yend = yend)) +
    geom_segment()

TidierPlots.geom_smooth — Method

geom_smooth(aes(...), ...)
geom_smooth(plot::GGPlot, aes(...), ...)

Represent data as a smoothed or linear fit.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Lines)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

NA

Optional Arguments

method: either "smooth" (default, loess fit) or "lm" (linear fit)

color / colour

linewidth

alpha

linestyle / linetype

Examples

xs = range(0, 2pi, length=30)
ys = sin.(xs) .+ randn(length(xs)) * 0.5
df = DataFrame(x = xs, y = ys)

ggplot(df, @aes(x = x, y = y)) + geom_smooth() + geom_point()

penguins = penguins_data();

ggplot(penguins, @aes(x = bill_length_mm, y = bill_depth_mm)) +
    geom_smooth(color=:red, linewidth=10, alpha=0.5)

TidierPlots.geom_step — Function

geom_step(aes(...), ...)
geom_step(plot::GGPlot, aes(...), ...)

Represents data as a stairstep plot.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Stairs)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

colormap / palette

linestyle / linetype

linewidth

alpha

Examples

xs = range(0, 2pi, length=30)
df = DataFrame(x = xs, y = sin.(xs))

ggplot(df, @aes(x = x, y = y)) + geom_step()

TidierPlots.geom_text — Function

geom_text(aes(...), ...)
geom_text(plot::GGPlot, aes(...), ...)

Plot text on a graph.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Text)

Required Aesthetics

x

y

text

Optional Aesthetics (see aes)

color / colour

Optional Arguments

color / colour

colormap / palette

align: tuple of positions (e.g. (:left, :bottom))

font

justification

rotation

fontsize

strokewidth

strokecolor

glowwidth / glow

glowcolor / glowcolour

word_wrap_width

alpha

Examples

df = DataFrame(
    x = [1,1,2,2],
    y = [1,2,1,2],
    t = ["A", "B", "C", "D"]
)

ggplot(df, @aes(x=x, y=y, text=t, color=t)) + geom_text()

ggplot(df, @aes(x=x, y=y, color=t)) +
    geom_text(@aes(text=t), fontsize=24, align=(:left, :bottom), font=:bold) +
    geom_point() +
    lims(x = (0, 3), y = (0, 3))

TidierPlots.geom_tile — Function

geom_tile(aes(...), ...)
geom_tile(plot::GGPlot, aes(...), ...)

Plots a heatmap as a collection of rectangles.

Details

x, y, and z must all be the same length, and there must be no duplicate (x, y) pairs. You can think of x, y, and z as triples of the form (x, y, f(x, y)).

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Heatmap)

Required Aesthetics

x

y

z

Optional Aesthetics (see aes)

NA

Optional Arguments

interpolate=false

colormap / palette

alpha

Examples

function mandelbrot(x, y)
    z = c = x + y*im
    for i in 1:30.0; abs(z) > 2 && return i; z = z^2 + c; end; 0
end

xs = -2:0.01:1
ys = -1.1:0.01:1.1
xys = Iterators.product(xs, ys) |> collect |> vec
zs = map(xy -> mandelbrot(xy[1], xy[2]), xys)

df = DataFrame(
    x = first.(xys),
    y = last.(xys),
    z = zs
)

ggplot(df, @aes(x = x, y = y, z = z)) + geom_tile()

TidierPlots.geom_violin — Function

geom_(aes(...), ...)
geom_(plot::GGPlot, aes(...), ...)

Represents data as a violin plot.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Violin)

Required Aesthetics

x (integer or categorical)

y (numeric)

Optional Aesthetics (see aes)

color / colour (used in conjunction with dodge)

dodge

Optional Arguments

orientation=:vertical: orientation of box (:vertical or :horizontal)

width=1

gap=0.2

show_notch=false

nothchwidth=0.5

show_median=true

dodge_gap=0.03

Examples

penguins = penguins_data();

ggplot(penguins, @aes(x=species, y=bill_length_mm)) +
    geom_violin()

ggplot(penguins, @aes(x=species, y=bill_length_mm)) +
    geom_violin(orientation=:horizontal)

ggplot(penguins, @aes(x=species, y=bill_length_mm, fill=sex, dodge=sex)) +
    geom_violin()

TidierPlots.geom_vline — Method

geom_vline(aes(...), ...)
geom_vline(plot::GGPlot, aes(...), ...)

Plot a horizontal line at the given y-intercept(s).

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.VLines)

Required Aesthetics

NA

Optional Aesthetics (see aes)

xintercept(s)

color / colour

Optional Arguments

xintercept(s)

color / colour

linewidth

linestyle / linetype

colormap / palette

alpha

Examples

# Plot only a single x-intercept
ggplot() + geom_vline(xintercept = 3)

# Plot multiple x-intercepts
ggplot() + geom_vline(xintercept = [-1, 4])

# Plot multiple x-intercepts mapped to a column
df = DataFrame(x = rand(4))
ggplot(df, @aes(xintercept = x)) + geom_vline()

TidierPlots.guides — Method

Sets which scales will get legends or colorbars. Use the name of the scale and either the string "legend" or "colorbar" to add a guide.

TidierPlots.label_bytes — Method

label_bytes(;units=:si, kwargs...)

Convert numeric values into human-readable strings representing byte sizes.

Arguments

units: Can be :si for SI units (powers of 10), :binary for binary units (powers of 1024), or a specific unit string (e.g., MB, GiB).

kwargs...: Additional keyword arguments passed to label_number.

Examples

labeler = label_bytes(units=:si)
labeler([1024, 2048]) # ["1.02 KB", "2.05 KB"]

TidierPlots.label_currency — Method

label_currency(;kwargs...)

Convert numeric values into currency strings.

Arguments

kwargs: Additional keyword arguments passed to label_number.

Examples

labeler = label_currency()
labeler([100, 200.75]) # ["$100", "$200.75"]

TidierPlots.label_date — Method

label_date(;format="m/d/Y")

Convert Date or DateTime values into formatted strings.

Arguments

format: A date format string.

Examples

labeler = label_date(format="Y-m-d")
labeler([Date(2020, 1, 1)]) # ["2020-1-1"]

TidierPlots.label_log — Method

label_log(;base=10, kwargs...)

Convert numeric values into logarithmic strings with a specified base.

Arguments

base: The logarithmic base.

kwargs: Additional keyword arguments passed to label_number.

Examples

labeler = label_log(base=10)
labeler([10, 100]) # ["10^1", "10^2"]

TidierPlots.label_number — Method

label_number(;precision=2, scale=1, prefix="", suffix="", decimal_mark=".", comma_mark=",", style_positive=:none, style_negative=:hyphen, kwargs...)

Format numeric values as strings with various styling options.

Arguments

precision: Number of decimal places.

scale: Scaling factor applied to the numbers before formatting.

prefix: String to prepend to the number.

suffix: String to append to the number.

decimal_mark: Character to use as the decimal point.

comma_mark: Character to use as the thousands separator.

style_positive: Style for positive numbers (:none, :plus, :space).

style_negative: Style for negative numbers (:hyphen, :parens).

kwargs: Additional keyword arguments passed on to format from Format.jl

Examples

labeler = label_number(precision=0, suffix="kg")
labeler([1500.12, -2000.12]) # ["1,500kg", "-2,000kg"]

TidierPlots.label_ordinal — Method

label_ordinal(;kwargs...)

Convert numeric values into ordinal strings (e.g., 1st, 2nd, 3rd).

Arguments

kwargs: Additional keyword arguments passed to label_number.

Examples

labeler = label_ordinal()
labeler([1, 2, 3]) # ["1st", "2nd", "3rd"]

TidierPlots.label_percent — Method

label_percent(;kwargs...)

Convert numeric values into percentage strings.

Arguments

kwargs: Additional keyword arguments passed to label_number.

Examples

labeler = label_percent()
labeler([0.1, 0.256]) # ["10%", "25.6%"]

TidierPlots.label_pvalue — Method

label_pvalue(;precision=2, kwargs...)

Format p-values, handling very small or large values with special notation.

Arguments

precision: Number of decimal places for thresholding small values.

kwargs: Additional keyword arguments passed to label_number.

Examples

labeler = label_pvalue()
labeler([0.0001, 0.05, 0.9999]) # ["<0.01", "0.05", ">0.99"]

TidierPlots.label_scientific — Method

label_scientific(;kwargs...)

Convert numeric values into scientific notation strings.

Arguments

kwargs: Additional keyword arguments passed to label_number.

Examples

labeler = label_scientific()
labeler([1000, 2000000]) # ["1e+03", "2e+06"]

TidierPlots.label_wrap — Method

label_wrap(width)

Wrap text strings to a specified width, breaking at spaces.

Arguments

width: The maximum number of characters in a line before wrapping.

Examples

labeler = label_wrap(10)
labeler(["This is a long sentence."]) # ["This is a
long
sentence."]

Reference - Internal functions

TidierPlots.alpha_scale_to_ggoptions — Method

Internal function. Converts args_dict to AxisOptions.

TidierPlots.handle_area — Method

geom_area(aes(...), ...)
geom_area(plot::GGPlot, aes(...), ...)

Displays a filled area from 0 to y.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Band)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour (outline color)

fill (fill color)

Optional Arguments

color / colour

fill

alpha

Examples

xs = range(0, 2pi, length=30)
df = DataFrame(x = xs, y = sin.(xs) .+ 1)

ggplot(df, @aes(x = x, y = y)) + geom_area()

TidierPlots.handle_point_color_and_fill — Method

Color and fill work slightly strangely in geom_point in ggplot2. Replicates behaviour.

In ggplot2:

fill = interior color of the point

color/colour = border (stroke) color of the point

In Makie Scatter:

color = interior color

strokecolor = border color

This function translates between the two conventions for both:

aes_dict: aesthetics mapped to data columns

args_dict: static values passed as arguments

TidierPlots.handle_polygon — Method

geom_polygon(aes(...), ...)
geom_polygon(plot::GGPlot, aes(...), ...)

Draws polygons defined by x and y coordinates.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Poly)

Required Aesthetics

x

y

Optional Aesthetics (see aes)

color / colour (outline color)

fill (fill color)

group (to define separate polygons)

Optional Arguments

color / colour

fill

strokewidth

alpha

Examples

# Single triangle
df = DataFrame(x = [0, 1, 0.5], y = [0, 0, 1])
ggplot(df, @aes(x = x, y = y)) + geom_polygon()

# Multiple polygons using group
df = DataFrame(
    x = [0, 1, 0.5, 2, 3, 2.5],
    y = [0, 0, 1, 0, 0, 1],
    group = [1, 1, 1, 2, 2, 2]
)
ggplot(df, @aes(x = x, y = y, group = group)) + geom_polygon()

TidierPlots.handle_rect — Method

geom_rect(aes(...), ...)
geom_rect(plot::GGPlot, aes(...), ...)

Draws rectangles defined by corner coordinates.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Poly)

Required Aesthetics

xmin

xmax

ymin

ymax

Optional Aesthetics (see aes)

color / colour (outline color)

fill (fill color)

Optional Arguments

color / colour

fill

strokewidth

alpha

Examples

df = DataFrame(
    xmin = [1, 3],
    xmax = [2, 4],
    ymin = [1, 2],
    ymax = [3, 4]
)

ggplot(df, @aes(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax)) + geom_rect()

TidierPlots.handle_ribbon — Method

geom_ribbon(aes(...), ...)
geom_ribbon(plot::GGPlot, aes(...), ...)

Displays a filled area between ymin and ymax.

Arguments

plot::GGPlot (optional): a plot object to add this geom to

aes(...): the names of the columns in the DataFrame that will be used in the mapping

...: options that are not mapped to a column (passed to Makie.Band)

Required Aesthetics

x

ymin

ymax

Optional Aesthetics (see aes)

color / colour (outline color)

fill (fill color)

Optional Arguments

color / colour

fill

alpha

Examples

xs = range(0, 2pi, length=30)
df = DataFrame(x = xs, lower = sin.(xs) .- 0.5, upper = sin.(xs) .+ 0.5)

ggplot(df, @aes(x = x, ymin = lower, ymax = upper)) + geom_ribbon()

TidierPlots.linewidth_scale_to_ggoptions — Method

Internal function. Converts args_dict to AxisOptions.

TidierPlots.make_alpha_lookup_continuous — Method

Internal function. Takes a Dict and makes a function that maps a numeric vector to the range specfied in the Dict's :range key, or to [0.1, 1.0] if not specified.

TidierPlots.make_linewidth_lookup_continuous — Method

Internal function. Takes a Dict and makes a function that maps a numeric vector to the range specfied in the Dict's :range key, or to [0.1, 1.0] if not specified.

TidierPlots.position_facets — Function

Internal function. Given a list of names and (optionally) some constraints, return the relative position of the facets and their labels.

TidierPlots.scale_custom — Function

scale_custom(aes, function, legend_options = Dict())
scale_custom(plot, aes, function, legend_options = Dict())

Create a custom scale for any aes.

Arguments

plot: Optional. GGPlot to apply this scale to

aes: The aes that the scale is for

f: Function to apply to the raw data. Should accept one argument (vector of data as it appears in your DataFrame) and return a vector in the format expected by Makie.

legend_options: Used to generate the guides for the plot

TidierPlots.shape_scale_to_ggoptions — Method

Internal function. Converts args_dict to AxisOptions.

TidierPlots.size_scale_to_ggoptions — Method

Internal function. Converts args_dict to AxisOptions.

TidierPlots.update_legend — Method

update_legend(legend, plot_axis_options, geom_name, a, labels_aes_df, 
              colorbar_highlim, colorbar_lowlim, colorbar_kwargs, colorbar)

Updates the legend or colorbar for a given aesthetic attribute.

Determines whether to display a colorbar or legend based on the legend options for the aesthetic. Handles special cases for different aesthetic types (markersize, alpha, linewidth, marker, color) and adds appropriate entries to the legend DataFrame. Automatically selects between colorbar and legend based on scale type (continuous/binned → colorbar, discrete/manual → legend).

Returns a tuple of updated (colorbarhighlim, colorbarlowlim, colorbar_kwargs, colorbar, legend).